Splitt ut utils-funksjoner i hver sin fil
