+++ src/frontend/Sider/Behandling/Inngangsvilkår/Målgruppe/EndreMålgruppeLæremidler.tsx
@@ -11,7 +11,7 @@ import {
     mapFaktaOgSvarTilRequest,
     nyMålgruppe,
     resettMålgruppe,
-} from './utils';
+} from './utilsMålgruppeLæremidler';
 import { MålgruppeValidering, validerMålgruppe } from './valideringMålgruppe';
 import { useBehandling } from '../../../../context/BehandlingContext';
 import { useInngangsvilkår } from '../../../../context/InngangsvilkårContext';
@@ -88,10 +88,17 @@ export const EndreMålgruppeLæremidler: React.FC<{
         form.type,
         form.vurderinger
     );
+    const begrunnelseErPåkrevd = delvilkårSomKreverBegrunnelse.length > 0;
+
     const kanEndreType = målgruppe === undefined && registerYtelsePeriode === undefined;
 
     const validerForm = (): boolean => {
-        const vilkårsperiodeFeil = validerMålgruppe(form, målgruppe, behandling.revurderFra);
+        const vilkårsperiodeFeil = validerMålgruppe(
+            form,
+            begrunnelseErPåkrevd,
+            målgruppe,
+            behandling.revurderFra
+        );
         settVilkårsperiodeFeil(vilkårsperiodeFeil);
 
         return isValid(vilkårsperiodeFeil);
